+++ lib/packages/core/src/api/__tests__/e2e/helpers/environment.ts
@@ -32,8 +32,8 @@ export const loadEnvironment = () => {
     testRecipientId:  process.env.TEST_RECIPIENT_ID,
     testRecipientPassphrase:  process.env.TEST_RECIPIENT_PASSPHRASE,
     testTransactionId:  process.env.TEST_TRANSACTION_ID,
-    testEncryptedMessageTransactionId: process.env.TEST_ENCRYPTED_MESSAGE_TRANSACTION_ID
-
+    testEncryptedMessageTransactionId: process.env.TEST_ENCRYPTED_MESSAGE_TRANSACTION_ID,
+    testContractId: process.env.TEST_CONTRACT_ID
     // .. add more variables here
   };